fix(logging): Enhance logging for here matching progress
Implement a more declarative usage of the logging facility, since it's the recommended way to use this package. It includes better readability, straightforward log delegation to the system logger, shared logging facility, sensible update interval instead of the modulo check and the removal of a random memory leak from the for-with implementation by using the builder class. It should be termination safe, but somehow isn't always. The progressbar is now configured to use the ConsoleProgressBarConsumer with a defined printStream. The printStream allows overwriting of prior lines with a carriage return and an uncomplicated configuration of the log level.