fix(logging): Implement a basic progressbar logging class
The class is designed to capture the current console pattern and output the progress bar in one line instead of multiple by keeping the style. Some very trivial tests are added to the custom class to ensure correct initialization. Once the progress is done, the bar automatically gets deleted from the log output.